Switch to ruby-fogbugz gem for API calls
This essentially removes everything related to curl usage and strips down the FogbugzService class; all related tests are removed too as this is now the responsibility of ruby-fogbugz.

Support for FogBugz checkins is more prominently disabled, however the basic code is left in. This currently doesn't work with GitLab 2.6 anyway as it lacks the filenames from the post hook. Once GitLab provides this feature, submission should be done via regular FogBugz 8 API instead of the custom cvsSubmit URL.