Update Rakefile and provide Gemfile, bump to 0.0.2
Note: ruby-fogbugz currently has a minor flaw and throws authentication exceptions with `throw` when it should use `raise`. The Rakefile depends on the version of rubygems.org, but if you use bundler you will use my fork with the small fix for now. It's no biggy but will lead to hard to interpret Ruby errors when authentication fails.